+++ src/main/java/com/diareat/diareat/auth/service/KakaoAuthService.java
@@ -3,8 +3,8 @@ package com.diareat.diareat.auth.service;
 import com.diareat.diareat.auth.component.KakaoUserInfo;
 import com.diareat.diareat.auth.dto.KakaoUserInfoResponse;
 import com.diareat.diareat.user.domain.User;
-import com.diareat.diareat.user.dto.CreateUserDto;
-import com.diareat.diareat.user.dto.JoinUserDto;
+import com.diareat.diareat.user.dto.request.CreateUserDto;
+import com.diareat.diareat.user.dto.request.JoinUserDto;
 import com.diareat.diareat.user.repository.UserRepository;
 import com.diareat.diareat.util.api.ResponseCode;
 import com.diareat.diareat.util.exception.UserException;
