--- src/main/java/com/diareat/diareat/user/dto/FollowUserDto.java
@@ -1,14 +1,0 @@
-package com.diareat.diareat.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class FollowUserDto {
-
-    private Long userId;
-    private Long followUserId;
-}
