+++ src/test/java/com/diareat/diareat/controller/UserControllerTest.java
@@ -4,7 +4,12 @@ import com.diareat.diareat.auth.service.KakaoAuthService;
 import com.diareat.diareat.user.controller.UserController;
 import com.diareat.diareat.user.domain.BaseNutrition;
 import com.diareat.diareat.user.domain.User;
-import com.diareat.diareat.user.dto.*;
+import com.diareat.diareat.user.dto.request.UpdateUserDto;
+import com.diareat.diareat.user.dto.request.UpdateUserNutritionDto;
+import com.diareat.diareat.user.dto.response.ResponseSearchUserDto;
+import com.diareat.diareat.user.dto.response.ResponseSimpleUserDto;
+import com.diareat.diareat.user.dto.response.ResponseUserDto;
+import com.diareat.diareat.user.dto.response.ResponseUserNutritionDto;
 import com.diareat.diareat.user.service.UserService;
 import com.diareat.diareat.util.api.ApiResponse;
 import com.diareat.diareat.util.api.ResponseCode;
