+++ src/test/java/com/diareat/diareat/service/FoodServiceTest.java
@@ -8,7 +8,7 @@ import com.diareat.diareat.food.repository.FoodRepository;
 import com.diareat.diareat.food.service.FoodService;
 import com.diareat.diareat.user.domain.BaseNutrition;
 import com.diareat.diareat.user.domain.User;
-import com.diareat.diareat.user.dto.ResponseRankUserDto;
+import com.diareat.diareat.user.dto.response.ResponseRankUserDto;
 import com.diareat.diareat.user.repository.FollowRepository;
 import com.diareat.diareat.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
