+++ src/test/java/com/diareat/diareat/service/UserServiceTest.java
@@ -3,7 +3,13 @@ package com.diareat.diareat.service;
 import com.diareat.diareat.user.domain.BaseNutrition;
 import com.diareat.diareat.user.domain.Follow;
 import com.diareat.diareat.user.domain.User;
-import com.diareat.diareat.user.dto.*;
+import com.diareat.diareat.user.dto.request.CreateUserDto;
+import com.diareat.diareat.user.dto.request.UpdateUserDto;
+import com.diareat.diareat.user.dto.request.UpdateUserNutritionDto;
+import com.diareat.diareat.user.dto.response.ResponseSearchUserDto;
+import com.diareat.diareat.user.dto.response.ResponseSimpleUserDto;
+import com.diareat.diareat.user.dto.response.ResponseUserDto;
+import com.diareat.diareat.user.dto.response.ResponseUserNutritionDto;
 import com.diareat.diareat.user.repository.FollowRepository;
 import com.diareat.diareat.user.repository.UserRepository;
 import com.diareat.diareat.user.service.UserService;
